refactor: move whatsapp reactions into platform
