+++ src/platforms/whatsapp/reactions.ts
